requirements.txt
autopep8==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
docxcompose==1.3.4
idna==3.3
lxml==4.8.0
msal==1.17.0
Office365-REST-Python-Client==2.3.11
pycodestyle==2.8.0
pycparser==2.21
PyJWT==2.3.0
python-docx==0.8.11
pytz==2021.3
requests==2.27.1
six==1.16.0
toml==0.10.2
urllib3==1.26.8
